be/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework-stubs==3.16.5
